Image uploader separation done
